xcovignore修正
