+++ EngineerResume/Utility/Debug/Git.swift
@@ -1,0 +1,13 @@
+#if DEBUG
+    import Foundation
+
+    struct Git {
+        static let commitHash: String = {
+            guard let bundleInfo = Bundle.main.infoDictionary else {
+                return .unknown
+            }
+
+            return bundleInfo["CommitHash"] as? String ?? .unknown
+        }()
+    }
+#endif
