--- EngineerResume/Utility/Git.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct Git {
-    static let commitHash: String = {
-        guard let bundleInfo = Bundle.main.infoDictionary else {
-            return .unknown
-        }
-
-        return bundleInfo["CommitHash"] as? String ?? .unknown
-    }()
-}
